+++ ft_printf/print_int.c
@@ -6,7 +6,7 @@
 /*   By: cdapurif <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/21 16:03:51 by cdapurif          #+#    #+#             */
-/*   Updated: 2019/11/24 17:10:43 by curtman          ###   ########.fr       */
+/*   Updated: 2019/11/26 16:14:55 by cdapurif         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	ft_print_int(t_struct *data, va_list args)
 {
-	int len;
-	int nbr;
+	long long	len;
+	long long	nbr;
 
-	nbr = va_arg(args, int);
+	nbr = (long long)va_arg(args, int);
 	len = (ft_nblen(nbr) > data->precision) ? ft_nblen(nbr) : data->precision;
 	if (data->precision > -1 && data->flag == 1)
 		data->flag = 0;
