+++ ft_printf/print_lowhex.c
@@ -6,7 +6,7 @@
 /*   By: cdapurif <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/22 18:26:10 by cdapurif          #+#    #+#             */
-/*   Updated: 2019/11/24 17:12:21 by curtman          ###   ########.fr       */
+/*   Updated: 2019/11/26 16:15:19 by cdapurif         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 void	ft_print_lowhex(t_struct *data, va_list args)
 {
-	int				len;
-	unsigned int	hex;
+	long long			len;
+	unsigned long long	hex;
 
-	hex = va_arg(args, unsigned int);
+	hex = (unsigned long long)va_arg(args, unsigned int);
 	len = ft_nblen_hex(hex);
 	len = (len > data->precision) ? len : data->precision;
 	if (data->precision > -1 && data->flag == 1)
